Move practitioner requests and tests to separate files
