+++ mocks/service_mock.go
@@ -45,10 +45,11 @@ func (mr *MockServiceMockRecorder) CreateInsolvencyResource(dao interface{}) *go
 }
 
 // CreatePractitionersResource mocks base method
-func (m *MockService) CreatePractitionersResource(dao []models.PractitionerResourceDao, transactionID string) error {
+func (m *MockService) CreatePractitionersResource(dao []models.PractitionerResourceDao, transactionID string) (error, int) {
 	ret := m.ctrl.Call(m, "CreatePractitionersResource", dao, transactionID)
 	ret0, _ := ret[0].(error)
-	return ret0
+	ret1, _ := ret[1].(int)
+	return ret0, ret1
 }
 
 // CreatePractitionersResource indicates an expected call of CreatePractitionersResource
